+++ app.py
@@ -23,7 +23,9 @@ import random
 from werkzeug.utils import secure_filename
 # Bleach to prevent cross-site scripting (XSS) attacks, possible when user is posting a comment
 import bleach
-from src.blueprints.posts_blueprint import router as posts_router, sanitize_html
+# from src.blueprints.posts_blueprint import router as posts_router, sanitize_html
+# from src.blueprints.login_blueprint import router as login_router
+from src.blueprints.profile_blueprint import router as profile_router
 
 # Pillow for image processing
 from PIL import Image
@@ -49,7 +51,9 @@ socketio = SocketIO(app, cors_allowed_origins='*')
 
 app.debug = True
 
-app.register_blueprint(posts_router)
+# app.register_blueprint(posts_router)
+# app.register_blueprint(login_router, bcrypt = bcrypt)
+app.register_blueprint(profile_router)
 
 
 # DB connection
@@ -355,15 +359,6 @@ def verify_login():
 
     return redirect('/login')
 
-
-# @app.get('/logout')
-# def logout_through_page():
-#     if not user_repository_singleton.is_logged_in():
-#         flash('Unable to logout because you are not logged in.', category='error')
-#         return redirect('/')
-#     user_repository_singleton.logout_user()
-#     return redirect('/login')
-
 @app.post('/logout')
 def logout():
     if not user_repository_singleton.is_logged_in():
@@ -372,23 +367,12 @@ def logout():
     user_repository_singleton.logout_user()
     return redirect('/login')
 
-@app.post('/profile/<string:username>/edit/delete')
-def delete(username):
-    if not user_repository_singleton.is_logged_in():
-        flash('Unable to delete account because you are not logged in.', category='error')
-        return redirect('/')
-    user_repository_singleton.remove_user(username)
-    user_repository_singleton.logout_user()
-    flash('Account deleted', category='success')
-    return redirect('/register')
-
 @app.get('/register')
 def register():
     if user_repository_singleton.is_logged_in():
         flash('You are already logged in. Logout to make a new account', category='error')
         return redirect('/')
     return render_template('register.html', user=session.get('user'))
-
 
 @app.post('/register')
 def create_user():
@@ -494,100 +478,243 @@ def password_reset(token):
 
     return redirect(f'/password_reset/{token}')
 
-# Create a get request for the profile page.
-@app.get('/profile/<string:username>')
-def profile(username: str):
+
+# @app.post('/profile/<string:username>/edit/delete')
+# def delete(username):
+#     if not user_repository_singleton.is_logged_in():
+#         flash('Unable to delete account because you are not logged in.', category='error')
+#         return redirect('/')
+#     user_repository_singleton.remove_user(username)
+#     user_repository_singleton.logout_user()
+#     flash('Account deleted', category='success')
+#     return redirect('/register')
+
+# # Create a get request for the profile page.
+# @app.get('/profile/<string:username>')
+# def profile(username: str):
+#     if not user_repository_singleton.is_logged_in():
+#         return redirect('/login')
+#         # user = user_repository_singleton.get_user_by_username(username)
+#         # if not user:
+#         #     abort(403)
+#         # profile_picture = url_for(
+#         #     'static', filename='profile_pics/' + user.profile_picture)
+
+#     user = user_repository_singleton.get_user_by_username(username)
+#     if not user:
+#         abort(403)
+
+#     logged_in_user = user_repository_singleton.get_user_by_user_id(user_repository_singleton.get_user_user_id())
+#     if not logged_in_user:
+#         abort(403)
+#     is_following = False
+#     if logged_in_user.is_following(user):
+#         is_following = True
+    
+#     posts = post_repository_singleton.get_user_posts(user.user_id)
+
+#     profile_picture = url_for(
+#         'static', filename='profile_pics/' + user.profile_picture)
+#     return render_template('profile.html', user=user, profile_picture=profile_picture, posts=posts, followers = user.get_all_followers(), following = user.get_all_following(), sanitize_html=sanitize_html, is_following=is_following)
+
+
+
+# @app.get('/profile/<string:username>/edit')
+# def get_edit_profile_page(username: str):
+#     if not user_repository_singleton.is_logged_in():
+#         abort(401)
+
+#     user_to_edit = users.query.filter_by(username=username).first()
+#     if user_to_edit is None:
+#         redirect(f'/profile/{username}')
+#     return render_template('edit_profile.html', user=user_to_edit)
+
+
+# @app.post('/profile/<string:username>')
+# def update_profile(username: str):
+#     if 'user' not in session:
+#         abort(401)
+
+#     user_to_edit = users.query.filter_by(username=username).first()
+
+#     if not user_to_edit:
+#         abort(403)
+
+#     new_email = request.form.get('email')
+#     new_username = request.form.get('username')
+#     new_fname = request.form.get('first_name')
+#     new_lname = request.form.get('last_name')
+
+#     existing_user = users.query.filter_by(username=new_username).first()
+#     existing_email = users.query.filter_by(email=new_email).first()
+
+#     if existing_user and existing_user != user_to_edit:
+#         flash('Username already in use', 'error')
+#         return redirect(f'/profile/{username}/edit')
+#     if existing_email and existing_email != user_to_edit:
+#         flash('Email already in use', 'error')
+#         return redirect(f'/profile/{username}/edit')
+
+#     profile_picture = request.files['image_upload']
+#     if profile_picture:
+#         filename = secure_filename(profile_picture.filename)
+#         if filename and allowed_file(filename):
+#         # Set UUID to prevent same file names
+#             pic_name = str(uuid.uuid1()) + "_" + filename
+#             profile_picture.save(os.path.join(
+#             app.config['UPLOAD_FOLDER'], pic_name))
+#             try:
+#                 img = Image.open(os.path.join(
+#                 app.config['UPLOAD_FOLDER'], pic_name))
+#                 img.verify()
+#             except Exception as e:
+#                 os.remove(os.path.join(app.config['UPLOAD_FOLDER'], pic_name))
+#                 abort(400, description="Upload file is not a valid image")
+#             user_to_edit.profile_picture = pic_name
+
+#     user_to_edit.email = new_email
+#     user_to_edit.username = new_username
+#     user_to_edit.first_name = new_fname
+#     user_to_edit.last_name = new_lname
+
+#     user_repository_singleton.login_user(user_to_edit)
+
+#     db.session.add(user_to_edit)
+#     db.session.commit()
+
+#     return redirect(f'/profile/{new_username}')
+
+@app.get('/posts/edit/<int:post_id>')
+def edit_post(post_id):
     if not user_repository_singleton.is_logged_in():
         return redirect('/login')
-        # user = user_repository_singleton.get_user_by_username(username)
-        # if not user:
-        #     abort(403)
-        # profile_picture = url_for(
-        #     'static', filename='profile_pics/' + user.profile_picture)
-
-    user = user_repository_singleton.get_user_by_username(username)
-    if not user:
-        abort(403)
-
-    logged_in_user = user_repository_singleton.get_user_by_user_id(user_repository_singleton.get_user_user_id())
-    if not logged_in_user:
-        abort(403)
-    is_following = False
-    if logged_in_user.is_following(user):
-        is_following = True
-    
-    posts = post_repository_singleton.get_user_posts(user.user_id)
-
-    profile_picture = url_for(
-        'static', filename='profile_pics/' + user.profile_picture)
-    return render_template('profile.html', user=user, profile_picture=profile_picture, posts=posts, followers = user.get_all_followers(), following = user.get_all_following(), sanitize_html=sanitize_html, is_following=is_following)
-
-
-
-@app.get('/profile/<string:username>/edit')
-def get_edit_profile_page(username: str):
-    if not user_repository_singleton.is_logged_in():
+    post = post_repository_singleton.get_post_by_id(post_id)
+    user = user_repository_singleton.get_user_by_username(
+        user_repository_singleton.get_user_username())
+    if not post or not user:
         abort(401)
-
-    user_to_edit = users.query.filter_by(username=username).first()
-    if user_to_edit is None:
-        redirect(f'/profile/{username}')
-    return render_template('edit_profile.html', user=user_to_edit)
-
-
-@app.post('/profile/<string:username>')
-def update_profile(username: str):
-    if 'user' not in session:
+    if post.user_id is not user.user_id:
         abort(401)
+    return render_template('edit_post.html', post=post, user=user)
 
-    user_to_edit = users.query.filter_by(username=username).first()
+# update a post
+@app.post('/posts/update/<int:post_id>')
+def update_post(post_id):
+    if post_id == '' or post_id is None:
+        abort(400)
+    post = post_repository_singleton.get_post_by_id(post_id)
+    user = user_repository_singleton.get_user_by_username(
+        user_repository_singleton.get_user_username())
+    if not post or not user:
+        abort(401)
+    if post.user_id is not user.user_id:
+        abort(401)
+    title = request.form.get('title')
+    description = request.form.get('text')
+    if title == '' or title is None:
+        abort(400)
 
-    if not user_to_edit:
-        abort(403)
+    image_upload = request.files.get('image_upload')
 
-    new_email = request.form.get('email')
-    new_username = request.form.get('username')
-    new_fname = request.form.get('first_name')
-    new_lname = request.form.get('last_name')
-
-    existing_user = users.query.filter_by(username=new_username).first()
-    existing_email = users.query.filter_by(email=new_email).first()
-
-    if existing_user and existing_user != user_to_edit:
-        flash('Username already in use', 'error')
-        return redirect(f'/profile/{username}/edit')
-    if existing_email and existing_email != user_to_edit:
-        flash('Email already in use', 'error')
-        return redirect(f'/profile/{username}/edit')
-
-    profile_picture = request.files['image_upload']
-    if profile_picture:
-        filename = secure_filename(profile_picture.filename)
+    if image_upload is not None:
+        filename = secure_filename(
+            image_upload.filename) if image_upload.filename else ''
         if filename and allowed_file(filename):
-        # Set UUID to prevent same file names
+            # Set UUID to prevent same file names
             pic_name = str(uuid.uuid1()) + "_" + filename
-            profile_picture.save(os.path.join(
-            app.config['UPLOAD_FOLDER'], pic_name))
+
+            # Save the file
+            image_upload.save(os.path.join(
+                app.config['POST_UPLOAD_FOLDER'], pic_name))
+
+            # Verify the file is an image using Pillow
             try:
                 img = Image.open(os.path.join(
-                app.config['UPLOAD_FOLDER'], pic_name))
-                img.verify()
+                    app.config['POST_UPLOAD_FOLDER'], pic_name))
+                img.verify()  # This will raise an exception if the file is not a valid image
             except Exception as e:
-                os.remove(os.path.join(app.config['UPLOAD_FOLDER'], pic_name))
-                abort(400, description="Upload file is not a valid image")
-            user_to_edit.profile_picture = pic_name
+                # Remove the invalid file
+                os.remove(os.path.join(
+                    app.config['POST_UPLOAD_FOLDER'], pic_name))
+                abort(400, description="Uploaded file is not a valid image.")
+            image_upload = pic_name
 
-    user_to_edit.email = new_email
-    user_to_edit.username = new_username
-    user_to_edit.first_name = new_fname
-    user_to_edit.last_name = new_lname
+    post_repository_singleton.update_post(
+        post_id, title, description, image_upload)
+    return redirect(f'/posts/{post_id}')
 
-    user_repository_singleton.login_user(user_to_edit)
+# delete a post
+@app.post('/posts/delete/<int:post_id>')
+def delete_post(post_id):
+    if post_id == '' or post_id is None:
+        abort(400)
+    post = post_repository_singleton.get_post_by_id(post_id)
+    user = user_repository_singleton.get_user_by_username(
+        user_repository_singleton.get_user_username())
+    if not post or not user:
+        abort(401)
+    if post.user_id is not user.user_id:
+        abort(401)
+    if post_repository_singleton.delete_post(post_id):
+        return redirect('/posts')
+    else:
+        abort(400)
 
-    db.session.add(user_to_edit)
-    db.session.commit()
+# when a user likes a post
+@app.post('/posts/like')
+def like_post():
+    post_id = request.form.get('post_id')
+    user_id = request.form.get('user_id')
+    if post_id == '' or post_id is None or user_id == '' or user_id is None:
+        abort(400)
+    post_repository_singleton.add_like(post_id, user_id)
 
-    return redirect(f'/profile/{new_username}')
+    return jsonify({'status': 'success'})
+
+# when a user likes a comment
+@app.post('/posts/like_comment')
+def like_comment():
+    comment_id = request.form.get('comment_id')
+    user_id = request.form.get('user_id')
+    if comment_id == '' or comment_id is None or user_id == '' or user_id is None:
+        abort(400)
+    post_repository_singleton.add_like_to_comment(comment_id, user_id)
+
+    return jsonify({'status': 'success'})
+
+# Function to sanitize HTML content
+def sanitize_html(content):
+    allowed_tags = ['p', 'div', 'em', 'strong', 'del', 'a', 'img', 'h1', 'h2',
+                    'h3', 'h4', 'h5', 'h6', 'blockquote', 'ul', 'ol', 'li', 'hr', 'br', 'pre']
+    allowed_attributes = {'*': ['class', 'style'], 'a': ['href', 'target']}
+
+    sanitized_content = bleach.clean(
+        content, tags=allowed_tags, attributes=allowed_attributes)
+    return sanitized_content
+
+# for comments and replies
+@app.post('/posts/<int:post_id>/comment')
+@app.post('/posts/<int:post_id>/comment/<int:parent_comment_id>')
+def comment_reply(post_id, parent_comment_id=0):
+    user = user_repository_singleton.get_user_by_username(
+        user_repository_singleton.get_user_username())
+    if user is None:
+        abort(401)
+    user_id = user.user_id
+    content = request.form.get('content')
+    reply = request.form.get('reply')
+    if reply is not None:
+        content = reply
+    if post_id == '' or post_id is None or content == '' or content is None:
+        abort(400)
+    if parent_comment_id == 0:
+        post_repository_singleton.add_comment(user_id, post_id, content)
+    else:
+        print('parent comment id: ', parent_comment_id)
+        post_repository_singleton.add_comment(
+            user_id, post_id, content, parent_comment_id)
+
+    return redirect(f'/posts/{post_id}')
 
 
 # sokcetIO to handle comments:
@@ -650,6 +777,51 @@ def handle_message(message):
         # emitting new comments:
         socketio.emit('message', {'user_id': user_id, 'content': content,
             'post_id': new_comment.post_id, 'username' : session['user']['username']})
+        
+@app.get('/posts')
+def posts():
+    if not user_repository_singleton.is_logged_in():
+        return redirect('/login')
+    all_posts = post_repository_singleton.get_all_posts_with_users()
+    user = user_repository_singleton.get_user_by_username(
+        user_repository_singleton.get_user_username())
+    if not user:
+        abort(401)
+    following_posts = post_repository_singleton.get_all_posts_of_followed_users(
+        user.user_id)
+
+    return render_template('posts.html', list_posts_active=True, all_posts=all_posts, user=user, sanitize_html=sanitize_html)
+
+# Create a get request for the posts page, following posts.
+@app.get('/posts/following')
+def following_posts():
+    if not user_repository_singleton.is_logged_in():
+        return redirect('/login')
+    user = user_repository_singleton.get_user_by_username(
+        user_repository_singleton.get_user_username())
+    if not user:
+        abort(401)
+    following_posts = post_repository_singleton.get_all_posts_of_followed_users(
+        user.user_id)
+
+    return render_template('posts.html', following_posts_active=True, following_posts=following_posts, user=user, sanitize_html=sanitize_html)
+
+# Create a get request for single post page.
+@app.get('/posts/<int:post_id>')
+def post(post_id):
+    if not user_repository_singleton.is_logged_in():
+        return redirect('/login')
+    post = post_repository_singleton.get_post_by_id(post_id)
+    user = user_repository_singleton.get_user_by_username(
+        user_repository_singleton.get_user_username())
+    if not post or not user:
+        abort(400)
+
+    following = False
+
+    if user.is_following(post.creator):
+        following = True
+    return render_template('single_post.html', post=post, user=user, sanitize_html=sanitize_html, following=following)
 
 
 @socketio.on('connect')
